VRrig Upkeep, Menu update, Music added(muted)
VRrigs lost some features in the past version so I put them back and updated the one in the main menu.
Main menu is now blacked out and has colliders.
Main menu now takes the player to Demo stage
Music is now added in the game but is currently muted till options are implemented